WEBCLIENT-244 #Fixed issues 243 and 244
WEBCLIENT-243 #Corrected dialog image
Now dialogs use the icon that has the exclaimation point inside a triangle.

WEBCLIENT-244 #Tutorial now properly resets when user clicks on "clear"
Upon Clearing a query, the tutorial state and UI on the Temporal Query Simple mode are now both reset so that when user goes into Simple Temporal Query again, the tutorial is in it's initial state (tr=urned off), and all the UIs are reset.